feat(friendModel):add error handler for dup request and invalid id params

+++ students/steven/Canchu/models/friendModel.js
@@ -32,8 +32,10 @@ module.exports = {
             }
             const sql_friend = 'INSERT INTO friendship (user_id, status, friend_id) VALUES (?,?,?)'
             db.query(sql_friend, [my_id,'pending',friend_id], (error, results) => {
-                if (error) {
-                    console.error('Database error:', error);
+                if (error.code === 'ER_DUP_ENTRY') {
+                    console.error('Duplicate Request Error', error.code);
+                    return res.status(403).json({ error: 'You have already send a request to this person.' });
+                } else if (error) {
                     return res.status(500).json({ error: 'Server error' });
                 } else {
                     const resultId = results.insertId
@@ -63,11 +65,13 @@ module.exports = {
                 console.error('Database error:', error);
                 return res.status(500).json({ error: 'Server error' });
             }
+            if(results[0] === undefined){
+                return res.status(403).json({ error: 'User ID not existed' });
+            }
             if(results[0].user_id !== self_id && results[0].friend_id !== self_id){
                 console.error("要求刪除Request的ID為",self_id,"，但能同意的只有ID為",results[0].user_id,"或ID為",results[0].friend_id,"的使用者，id為",id)
-		        return res.status(400).json({ error: 'This user has no permission to agree friend request'})
+		        return res.status(400).json({ error: 'This user has no permission to delete friend request'})
             } else {
-                console.log("通過。接下來...")
                 const sql = 'DELETE FROM friendship WHERE id = ?'
                 db.query(sql, [id], (error, results) => {
                     if (error) {
@@ -91,8 +95,11 @@ module.exports = {
             if (error) {
                 console.error('Database error:', error);
                 return res.status(500).json({ error: 'Server error' });
-	    }
-	    console.log("印出來：",id,results,results[0])
+	        }
+	        console.log("印出來：",id,results,results[0])
+            if(results[0] === undefined){
+                return res.status(403).json({ error: 'User ID not existed' });
+            }
             if(results[0].user_id !== self_id){
                 console.error("要求同意Request的ID為",self_id,"，但能同意的只有ID為",results[0].user_id,"的使用者，id為",id)
 		        return res.status(400).json({ error: 'This user has no permission to agree friend request'})
